OK hourly wage peeps still don't get bonuses, also overtime not addressed at all. But I think for DIP at the moment this is good.
Overtime would require a weekly break down as far as I know. anything over 40 a week is overtime. but you can't just take 40*52 and call anything hire overtime cause 42 hour week is 2 hours overtime, 32 hour week is no over time but it does not affect the previous 42 hour week. This would require a weekly break down of hours for hourly employees. Which would require like 52 inputs. 

Possible solution to bonuses extend HRly emps to another class and give them bonuses and then overide the getAnnualwages you have to override it as an interface anyways... so that should be good. Even though it seems to fly in the face of inheritance to override something you are inheriting. But if you create a new class and copy paste in the HR class and then add and change it you are still overriding the interface. 

pretty much the same thing. Do you inherit the interface I wonder...  I bet you do. which means... it doesn't matter.

untill you change the program and add something to HrEmp that HrPlus doesn't need or want. but what could you add to HrEmp that HrPlus wouldn't need.